MR : Feat, Fix | hong => controller

See merge request s09-webmobile3-sub2/S09P12A304!134

+++ BackEnd/ggong/src/main/java/com/a304/ggong/dto/response/AnswerDetailResponse.java
@@ -16,7 +16,7 @@ public class AnswerDetailResponse {
 	private String dataLabel;
 	private Long answerA = 0L;
 	private Long answerB = 0L;
-	private double rateA;
-	private double rateB;
+	private Long rateA;
+	private Long rateB;
 
 }
